Update: worldwide.views 모델화로 분리 -DB등록코드도 마무리
